requirements.txt
anyio==3.6.2
APScheduler==3.6.3
asttokens==2.2.1
backcall==0.2.0
cachetools==4.2.2
certifi==2022.12.7
colorama==0.4.6
decorator==5.1.1
executing==1.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
ipython==8.8.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.36
pure-eval==0.2.2
Pygments==2.14.0
python-telegram-bot==13.14
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
scapy==2.5.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
tornado==6.1
traitlets==5.8.1
tzdata==2022.7
tzlocal==4.2
wcwidth==0.2.6
